upgrade to .net core sdk 2.1.300 (includes .net core 2.1)

+++ examples/csharp/run.ps1
@@ -3,21 +3,23 @@
 $env:DOTNET_CLI_TELEMETRY_OPTOUT = '1'
 
 # install the dotnet sdk.
-# see https://github.com/dotnet/core/blob/master/release-notes/download-archives/2.1.103-sdk-download.md
-$cliVersion = '2.1.103'
-$cliHome = "c:\ProgramData\dotnet-sdk-$cliVersion"
-$archiveName = "dotnet-sdk-$cliVersion-win-x64.zip"
+# see https://github.com/dotnet/core/blob/master/release-notes/download-archives/2.1.0-download.md
+$archiveUrl = 'https://download.microsoft.com/download/8/8/5/88544F33-836A-49A5-8B67-451C24709A8F/dotnet-sdk-2.1.300-win-x64.exe'
+$archiveHash = 'f01012683591f6a9530b27296c287390e9a88cad6cb20339366e3ff627f25e26965ab9bd9f982d8010c4d195943c18f7d5e08e246d315d15e48aaaf07cd5a372'
+$archiveName = Split-Path -Leaf $archiveUrl
 $archivePath = "$env:TEMP\$archiveName"
 Write-Host "Downloading $archiveName..."
-Invoke-WebRequest "https://download.microsoft.com/download/E/2/6/E266C257-F7AF-4E79-8EA2-DF26031C84E2/$archiveName" -UseBasicParsing -OutFile $archivePath
-Expand-Archive $archivePath -DestinationPath $cliHome
+(New-Object Net.WebClient).DownloadFile($archiveUrl, $archivePath)
+$archiveActualHash = (Get-FileHash $archivePath -Algorithm SHA512).Hash
+if ($archiveHash -ne $archiveActualHash) {
+    throw "$archiveName downloaded from $archiveUrl to $archivePath has $archiveActualHash hash witch does not match the expected $archiveHash"
+}
+Write-Host "Installing $archiveName..."
+&$archivePath /install /quiet /norestart | Out-String -Stream
+if ($LASTEXITCODE) {
+    throw "Failed to install dotnetcore-sdk with Exit Code $LASTEXITCODE"
+}
 Remove-Item $archivePath
-
-# add dotnet to the Machine PATH.
-[Environment]::SetEnvironmentVariable(
-    'PATH',
-    "$([Environment]::GetEnvironmentVariable('PATH', 'Machine'));$cliHome",
-    'Machine')
 
 # update $env:PATH with the recently installed Chocolatey packages.
 Import-Module C:\ProgramData\chocolatey\helpers\chocolateyInstaller.psm1
